Add models related to user
