+++ backend/rumergy_backend/rumergy/models/access_request.py
@@ -1,0 +1,5 @@
+from django.db import models
+from . import User
+
+class AccessRequest(models.Model):
+    user = models.ForeignKey(User, on_delete=models.PROTECT)
