+++ backend/rumergy_backend/rumergy/models/data_log.py
@@ -1,0 +1,4 @@
+from django.db import models
+
+class DataLog(models.Model):
+    pass
