+++ backend/rumergy_backend/rumergy/models/user.py
@@ -1,9 +1,48 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-from django.contrib.auth.models import AbstractBaseUser
-from django.contrib.auth.models import PermissionsMixin
-from django.contrib.auth.models import BaseUserManager
-from .user_profile_manager import UserProfileManager
+from django.contrib.auth.models import (
+    AbstractBaseUser,
+    PermissionsMixin,
+    BaseUserManager,
+)
+
+
+class UserProfileManager(BaseUserManager):
+    """Manager for user profiles"""
+
+    def create_user(self, email, first_name, last_name, role, password=None):
+        """Create a new user profile"""
+        if not email:
+            raise ValueError("User must have an email address")
+        if not first_name:
+            raise ValueError("User must have a first name")
+        if not last_name:
+            raise ValueError("User must have a last name")
+        if not role:
+            raise ValueError("User must have a role")
+
+        email = self.normalize_email(email)
+        user = self.model(
+            email=email,
+            first_name=first_name,
+            last_name=last_name,
+            role=role,
+        )
+
+        user.set_password(password)
+        user.save(using=self._db)
+
+        return user
+
+    def create_superuser(self, email, first_name, last_name, role, password=None):
+        """Create a new superuser profile"""
+        user = self.create_user(email, first_name, last_name, role, password)
+        user.is_superuser = True
+        user.is_staff = True
+
+        user.save(using=self._db)
+
+        return user
 
 
 class User(AbstractBaseUser, PermissionsMixin):
